refactor(api): extract UPDATE_TEAM to handlers/team.rs
Step 3 Slice C1 — follows HandlerCtx pattern from Slice A.

New file: src/api/handlers/team.rs
- handle_update_team(params, ctx) → Value

handle_session changes:
- UPDATE_TEAM arm → one-liner call
- Remaining 2 arms: CREATE_TEAM + SHUTDOWN

## Refactor Pattern Change
UPDATE_TEAM missing-name error converted from writeln+continue to
return Value. All other logic is unchanged (team diff computation,
notifier event emission).

Ref: Slice A Pattern Established (b8d5c40)

Tests: 617 pass, 0 fail, 2 ignored. Characterization tests from
Commit 1 verify no behavior change.
